refactors loading of tokens so tokens.env can be placed in the parent directory and adds the chunking tests to the workflow
