+++ .gitignore
@@ -272,3 +272,7 @@ fabric.properties
 
 # Android studio 3.1+ serialized cache file
 .idea/caches/build_file_checksums.ser
+
+tokens.env
+QAChat/Data_Processing/client_secrets.json
+QAChat/Data_Processing/credentials.json
